requirements.txt
# Telegram ClawdBot - Python Dependencies

# Telegram Bot Framework
python-telegram-bot[job-queue]>=21.0

# AI/LLM
openai>=1.0

# Memory System
mem0ai>=0.1

# Configuration
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0

# Scheduling
apscheduler>=3.10

# Vector Operations
numpy>=1.24

# Logging (colorized output)
colorlog>=6.0
